feat(service): add ListTask with cache invalidation and 30 unit tests
- implement ListTask with cache-first strategy and cache invalidation on write ops
- add DeleteByPrefix to CacheStore interface and redis adapter (SCAN + UNLINK)
- add IncTaskListed/RecordTaskListedDuration metrics
- change PaginationResponse.Total from int to int64
- update all existing tests to verify DeleteByPrefix calls on success paths
- add 6 ListTask tests: cache hit, cache miss, status filter, empty list, db error, cache failure

+++ param/pagination.go
@@ -11,9 +11,9 @@ type PaginationRequest struct {
 }
 
 type PaginationResponse struct {
-	PageSize   int `json:"page_size"`
-	PageNumber int `json:"page_number"`
-	Total      int `json:"total"`
+	PageSize   int   `json:"page_size"`
+	PageNumber int   `json:"page_number"`
+	Total      int64 `json:"total"`
 }
 
 func (p *PaginationRequest) GetPageNumber() int {
